fix: moved custom interfaces out of generated type file
