+++ apps/tailwind-components/app/components/cms/ConfigurablePage.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import type { IConfigurablePages } from "../../../types/cms";
 import type { ITableMetaData } from "../../../../metadata-utils/src";
-import type { IDraggingInfo } from "../../../types/cms";
+import type { IDraggingInfo } from "../../../types/CmsComponents";
 
 import PageComponent from "./PageComponent.vue";
 import TextParagraph from "./Paragraph.vue";
